+++ Assets/Code/test/Editor/Match/Rules/TimeStockMatchRuleTest.cs
@@ -11,7 +11,7 @@ using NUnit.Framework;
 public class TimeStockMatchRuleTest {
 
   MatchState CreateGameState(uint time, int[] stocks) {
-    var states = stocks.Select(s => new PlayerState { Stocks = s });
+    var states = stocks.Select(s => new PlayerState { Stocks = (sbyte)s });
     return new MatchState(states) { Time = time };
   }
